+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 * :white_check_mark: Prefect Paintx testing
 * :white_check_mark: Prefect Colorx testing
 * :white_check_mark: Prefect Imagex testing
+* :fire: Drawablex toDrawableByColor rename to changeColor, toDrawableByColorFromDrawableRes rename to changeResDrawableColor
 
 Upgrade
 * :arrow_up: Upgrade Javax 0.9.0